Tradotto commenti in inglese e risolto qualche warning di Lint

+++ app/src/main/java/com/android/notificationexpo/BubbleActivity.kt
@@ -42,24 +42,19 @@ class BubbleActivity: AppCompatActivity() {
         //mostro la action bar
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
 
-        // savedInstanceState is non-null when there is fragment state
-        // saved from previous configurations of this activity
-        // (e.g. when rotating the screen from portrait to landscape).
-        // In this case, the fragment will automatically be re-added
-        // to its container so we don"t need to manually add it.
-        // For more information, see the Fragments API guide at:
-        //
-        // http://developer.android.com/guide/components/fragments.html
-        //
+        /*savedInstanceState è non null quando c'è un fragment già creato da una configurazione precedente di ItemDetailActivity
+      * (per esempio quando si ruota il dispositivo). In questo caso il fragment verrà automaticamente riaggiunto nel
+      *  FrameLayout (con id item_detail_container) perciò non è nessario aggiungerlo manualmente*/
+
         if (savedInstanceState == null) {
-            // Create the detail fragment and add it to the activity
-            // using a fragment transaction.
+            // Creo il fragment del dettaglio e lo aggiungo all'Activity
+            // usando una fragment transaction.
             val fragment = ItemDetailFragment().apply {
                 arguments = Bundle().apply {
                     //passo al ItemDetailFragment le informazioni di cui ha bisogno
                     putLong(ItemDetailFragment.CHAT_ID, intent.getLongExtra(ItemDetailFragment.CHAT_ID,-1))
                     putString(ItemDetailFragment.CHAT_NAME, intent.getStringExtra(ItemDetailFragment.CHAT_NAME))
-                    putInt(ItemDetailFragment.CHAT_IMG,intent.getIntExtra(ItemDetailFragment.CHAT_IMG, -1) as Int)
+                    putInt(ItemDetailFragment.CHAT_IMG, intent.getIntExtra(ItemDetailFragment.CHAT_IMG, -1))
                     putString(ItemDetailFragment.NOTIFICATION,intent.getStringExtra(ItemDetailFragment.NOTIFICATION))
                     putBoolean(ItemDetailFragment.TWO_PANE,intent.getBooleanExtra(ItemDetailFragment.TWO_PANE,false))
                 }
@@ -74,12 +69,7 @@ class BubbleActivity: AppCompatActivity() {
     override fun onOptionsItemSelected(item: MenuItem) =
             when (item.itemId) {
                 android.R.id.home -> {
-                    // This ID represents the Home or Up button. In the case of this
-                    // activity, the Up button is shown. For
-                    // more details, see the Navigation pattern on Android Design:
-                    //
-                    // http://developer.android.com/design/patterns/navigation.html#up-vs-back
-
+                    //Questo ID rappresenta il pulsante Home
                     navigateUpTo(Intent(this, ItemListActivity::class.java))
                     true
                 }
